triton_machine: Introduce locality stanza to support locality hints/affinity rules
* Add locality stanza to triton_machine schema
* Documentation for new locality stanza
* Acceptance tests for new locality stanza

This feature does not include the `strict` argument for locality hints within
Triton. This behavior returns errors if locality hints cannot be applied,
leading to issues with some architectures using Terraform. A future update to
the provider might find a more beneficial manner in which to utilize this but
it's not included today.