resource/triton_machine: Fix issues with machine network diffs
Fixes: #30

There was an issue that when you specified a pool as a ni network id,
the provisioned network id would be different to that supplied to the
machine and thus create a perpetual diff

We decided that networks for the machine to be part of and nics attached
to the machine are different enough for them to be separate. Therefore,
we un-deprecated `networks` and we now pass a list of network ids for
the machine to be attached to

`nic` is now used as a computed parameter rather than an input. Meaning
we can do the following:

```
provider "triton" {}

data "triton_image" "image" {
    name = "base-64-lts"
    version = "16.4.1"
}

data "triton_network" "public" {
    name = "Joyent-SDC-Public"
}

resource "triton_vlan" "test" {
	  vlan_id = 100
	  name = "test-vlan"
	  description = "test vlan"
}

resource "triton_fabric" "test" {
	name = "test-network"
	description = "test network"
	vlan_id = "${triton_vlan.test.vlan_id}"

	subnet = "10.10.0.0/24"
	gateway = "10.10.0.1"
	provision_start_ip = "10.10.0.10"
	provision_end_ip = "10.10.0.250"

	resolvers = ["8.8.8.8", "8.8.4.4"]
}

resource "triton_machine" "test" {
    package = "g4-highcpu-128M"
    image   = "${data.triton_image.image.id}"
    networks = ["${data.triton_network.public.id}", "${triton_fabric.test.id}"]
}
```

Much simplier setup of `triton_machine` and doesn't give the perpetual diff

```
% make testacc TEST=./triton TESTARGS='-run=TestAccTritonMachine_'                                                                                                              ✹
==> Checking that code complies with gofmt requirements...
TF_ACC=1 go test ./triton -v -run=TestAccTritonMachine_ -timeout 120m
=== RUN   TestAccTritonMachine_basic
--- PASS: TestAccTritonMachine_basic (91.62s)
=== RUN   TestAccTritonMachine_dns
--- PASS: TestAccTritonMachine_dns (225.36s)
=== RUN   TestAccTritonMachine_nic
--- PASS: TestAccTritonMachine_nic (174.16s)
=== RUN   TestAccTritonMachine_addNIC
--- PASS: TestAccTritonMachine_addNIC (410.63s)
=== RUN   TestAccTritonMachine_firewall
--- PASS: TestAccTritonMachine_firewall (348.12s)
=== RUN   TestAccTritonMachine_metadata
--- PASS: TestAccTritonMachine_metadata (294.64s)
=== RUN   TestAccTritonMachine_cns
--- PASS: TestAccTritonMachine_cns (274.49s)
=== RUN   TestAccTritonMachine_locality
--- PASS: TestAccTritonMachine_locality (441.67s)
PASS
ok	github.com/terraform-providers/terraform-provider-triton/triton	1851.249s
```