Add new `triton_fabric_vlan` data source.
This commit adds new data source called `triton_fabric_vlan`, which allows the
end-user to find details about a particular Fabric VLAN based on attributes such
as name, VLAN ID or description. Each of the attribute can be used as a filter
allowing to narrow down results more granularly (each filter creates an
implicit AND relationship with the other).

Additionally, this commit introduces a new `wildcardMatch` function that can be
used to perform simple pattern matches against a string value, and a function
`validateVLANIdentifier` that can be used in the schema where the `ValidateFunc`
is to be expected to verify whether the value for the Fabric VLAN ID is within
the correct range.

Signed-off-by: Krzysztof Wilczynski <kw@linux.com>